+++ apps/backoffice-v2/src/pages/SignIn/SignIn.page.tsx
@@ -56,10 +56,13 @@ export const SignIn: FunctionComponent = () => {
     },
   });
 
+  // Handles a flash of content on sign in
+  if (isAuthenticated) return null;
+
   return (
     <section className={`flex h-full flex-col items-center justify-center`}>
       <div className={`mb-16`}>
-        {env.VITE_IMAGE_LOGO_URL ? (
+        {!!env.VITE_IMAGE_LOGO_URL ? (
           <img className={`w-40`} src={env.VITE_IMAGE_LOGO_URL} />
         ) : (
           <BallerineLogo />