+++ packages/common/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ballerine/common
 
+## 0.5.3
+
+### Patch Changes
+
+- Update Common
+
+## 0.5.2
+
+### Patch Changes
+
+- New version
+
 ## 0.5.1
 
 ### Patch Changes