+++ services/workflows-service/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ballerine/workflows-service
 
+## 0.4.14
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.5.3
+
+## 0.4.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.5.2
+
 ## 0.4.12
 
 ### Patch Changes